+++ packages/table/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.1](https://github.com/tecsinapse/ui-kit/compare/@tecsinapse/table@3.5.0...@tecsinapse/table@3.5.1) (2021-03-03)
+
+**Note:** Version bump only for package @tecsinapse/table
+
+
+
+
+
+# [3.5.0](https://github.com/tecsinapse/ui-kit/compare/@tecsinapse/table@3.4.5...@tecsinapse/table@3.5.0) (2021-02-26)
+
+**Note:** Version bump only for package @tecsinapse/table
+
+
+
+
+
 ## [3.4.5](https://github.com/tecsinapse/ui-kit/compare/@tecsinapse/table@3.4.4...@tecsinapse/table@3.4.5) (2021-01-14)